classes/04-llm-applications-structured-output/homework/requirements.txt
openai
python-dotenv
colorama
jupyter
